+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added IEM parsed ASOS/AWOS station observation data source (`IEM_ASOS`)
 - Added Zarr v3 sharding support to `AsyncZarrBackend`
+- Added EarthMover NOAA GFS forecast data source (`EarthMoverGFS_FX`)
+- Added EarthMover NOAA GEFS analysis data source (`EarthMoverGEFS`)
+- Added EarthMover NOAA GEFS 35-day ensemble forecast data source (`EarthMoverGEFS_FX`)
 
 ### Changed